requirements.txt
pip==1.4.1
Django==1.8.5
Markdown==2.6.2
PyJWT==1.4.0
PyYAML==3.10
bbp-services==0.0.8
bleach==1.4.1
django-bower==5.0.4
django-formtools==1.0
django-jsonify==0.3.0
django-sslserver==0.15
django-widget-tweaks==1.4.1
hbp-app-python-auth==0.0.4
html5lib==1.0b8
oauthlib==1.0.3
python-openid==2.2.5
python-social-auth==0.2.0
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
wheel==0.24.0
wsgiref==0.1.2
